Fix corruption of outgoing_moves in packed state of stock.shipment.out
The `move_obj.write` changes the value of `quantity` in the move, but the
removed_qty computation used `move.quantity` from the browse record. The
browse record gets the updated value and not the old value from
browse_record cache. This fix just moves the computation before the
move_obj.write and includes a scenario to confirm the behavior.

issue2426
review254001