Fix reload of screen at the end of wizard
A wizard could be launch by a button defined on a children record. In this case,
the ids to reload is not the ids from the wizard as they are from the children
model but just the current record.

issue3519
review2571002
(transplanted from 6f657eeefffcc5a74be70e7cc8868fc7293d3c44)