Add wizard to reconcile accounts
Currently, the "Open Reconcile Lines" can become very difficult to use when you
have a lot of different parties and/or a lot of account to reconcile.
Here is a new process which is based on a wizard that loops over all the couple
of account/party that potentially have lines to reconcile. It shows for each
couple a proposal of reconciliation that the user can accept or modify (with
write-off possibilities).
The proposal is found by testing all the combinations of lines up to 10 lines by
default (configurable). As such problem is NP-hard, I think we can afford to use
the brute force for small number and let the user solves higher number or
iterate on running the wizard. Any way, it is possible to implement other
strategy by module extention.

issue4185
review14511002