Use 'set' attribute to decide column for table_query
Since changeset 64681f657d55, Function fields have a sql_type so it is wrong to
rely on this property to use or not the field. The 'set' method of field is
already used in many places for this purpose.

issue4434
(grafted from e10d64aec56cae7fa0452cd46abfc0e4246104e7)

--HG--
branch : 3.4