Track the amount invoiced in sales outside Europe
This data is needed on the VAT declaration. So instead of dropping
the tax in the "Ventes hors communauté" tax rule, we map them to
a new tax with a 0% rate which records the base amount in the correct
tax code (0032 for the invoice, 0039 for the credit note).

issue4569
review10961002