Remove tax rule "Ventes intracommunautaires" and associated taxes
The set of taxes used by this rule tried to cover EU-sales to individuals
and did a bad job of it.

This topic is complicated with lots of special cases (depending on
the gross sales, on the buyer's country, on the kind of sales, etc.)
and is thus not covered by the default rules.

Furthermore for the simplest cases, such sales must be accounted much
like domestic sales for which no special rule is needed.

review14011003