Create taxes for purchase of services in Europe
Basically we copy the taxes for purchases of goods, changing the tax codes
as appropriate.

The base amount is now recorded in 0044 instead of 0031, and the tax
amount doesn't need to be counted separately (there's no
service-equivalent of the 0035 field). This enables us to do everything
with 2 sub-taxes only.

review6081002