Fix TypeError during min/max computation in Graph
If the field is a datetime, a date or a timedelta but the value if None, it
could be compared to an existing instance and so raise a TypeError.
So the min/max value must be set with a correct value and TypeError must be
catched if the record value is None.

issue4888
review16501002