Change treeview overflow on completion shown
When the dropdown is shown in a treeview, if there is no enough space in the
treeview, the browser adds a scrollbar to the treeview instead of expanding the
height of the treeview (and let the general scrollbar managed it). This is
because the treeview has an auto overflow-x which lead to set also the same to
overflow-y.