Record.defaut_get must return a promise which resolves with values
The promise doesn't always return the values if there is any autocomplete and
the promise is resolved before the set_default is called.
So we use our custom deferred that will be resolved after the call to
set_default and the autocomplete promises.

issue5222
review19721002
(grafted from fad9c613411a3717a0bbdf832345d8808afcc017)

--HG--
branch : 3.8