Better format for month and week in reports
It is better to manage the formatting of those field on the server side to
ensure it doesn't depend on any local number formatting.
So we extract into an internal Char field the values (to ease searcher code)
and we format the using Python format via a Function field.

issue5274
review20691002