Remove Party Language configuration
The Party language is now a property for which the default value is no more
managed by default method.
Also it is probably better to leave the field empty if user doesn't know than
having a default value set like that a choice can be done at run time.

issue5198
review20111002