Allow request without product
The purchase line does not required to have a product, so the purchase request
should behave the same. A description field is added to be used instead of the
product name as purchase line description.
The unit price is only computed if it is not yet set and if there is a product.

issue5670
review27431002