Apply invoice's account clause only when the account is set
The account domain on invoice field is applied when the party is set, but it should only be applied when the account is set. This gives an invalid domain when the party is set and the account is not set.

issue5766
review32391002
(grafted from 686f607b0b3e0b9f1b2a8ffa841495c0cc8f8025)

--HG--
branch : 3.8