Add Many2Many targets to all records
If at least one record had already the target, the target was not added to any
records. This is because the test for existing link was only done on the target
id instead of the tuple origin and target.

issue5792
review26551002
(grafted from bd1ce2053e459b124344ef9801b9e9f12f728a47)

--HG--
branch : 3.6