Load autocomplete only on display of Char widget
Keeping the autocomplete always up to date is very expensive especially for
large list because the method is called once per field for each record.
Indeed autocomplete is only needed when the Char widget is displayed. So we
could just load it at display and keep it up to date when fields are changed.

issue6167
review31871002