Skip validate only on constant readonly fields
A field can be readonly by the domain inversion in the validate method. So it
should be validated again in case its domain inversion change and reset it a
non-readonly.
So only field that are always readonly can be skipped for the validation.

issue6448
review38341002
(grafted from e9613d7e20c4f3d793d949eccdaa82d66e80953c)

--HG--
branch : 3.8