Correctly set focus on filter form for Selection and Time fields
We can not rely on setting the focus on the entry as the selection and time
widgets store the input as separate element. So the focus must be set directly
to the input to guarantee that it works for all widgets.

issue6455
review34091002
(grafted from b8be0b47d84cf833caf0f7dc4f3bfb4ff93018e1)

--HG--
branch : 4.0