Compute average cost with right product quantity
The product quantity must be computed on all storage location but without
childs otherwise the children quantity are added a second time on parent by
Move._get_quantity
To allow this behaviour the Move._get_quantity and Move._search_quantity use
the 'with_childs' context. If the context is not set, the default value will
be False if there are more than one location. This provides a sane default
behaviour because the locations could be children from each others.

issue6627
review35491002