Add searcher on General Ledger
It is implemented by evaluating the field for all accounts because the
computation can not be reduced into a SQL query as it may require recursive
computation for all previous fiscal year.
It should not be a problem as the number of account is normally quite static
overtime and the number should not exceed the query size of the database
backend.

issue5168
review41521002