Use input number and toLocaleString for numeric fields
To get a proper virtual keyboard on mobile, it is needed to use a input of
type 'number' when editing a numeric field. But as it is not possible to
ensure to display non signifiant digits, a text input is used with the
toLocaleString method.
Also the lang attribute is repeated on the number input because not all
browsers manage to use the global as default.

issue5233
review43621002