Move income_statement flag to "Compte de résultat" The account type "Résultat de l'exercice [bénéfice ou perte]" has real accounts associated to it and due to this the income statement does not show the desired value on the top-level entry in the tree.
In particular when you close the account of year X, you write a benefit
(or loss) in "Résultat de l'exercice [bénéfice ou perte]" date at
31-12-X. But when your company holds its general assembly, it will
decide what to do with this money. Either report it back for re-use in
year X+1, or distribute it to the shareholders, or put it in a (legal)
reserve. This decision is then transcribed in the the accounting at the
date of the general assembly (which is in year X+1).

The money is moved out of 12X (Résultat) into either 11X (Report), 106
(Réserves), 457 (Associés - Dividendes à payer). And it is this move
which is polluting the income statement.

issue7070
review37931002