Remove all keyBinds from datetimepicker
datetimepicker stops the propagation and prevents default for event with key
bound even if the handler does nothing.
This prevents to have Return to trigger the default submit or Escape to stop
list edition.
But as datetimepicker uses jQuery.extend to setup the options, we can pass
null for keyBinds to remove all the defaults.

issue7246
review45301002