Re-use employee instance instead of reading from user
Get the employee from the user after having set it as current user return None
because the context of the instance still has the context of the admin user
(with employee=None) but applies to the user.
This is because set_user change globally the user and context of the config.

issue7144
review44831002