Re-use employee instance and append new Party instance
Get the employee from the user after having set it as current user return None
because the context of the instance still has the context of the admin user
(with employee=None) but applies to the user. This is because set_user change
globally the user and context of the config.
Appending instance to a many2many require an instance with the same context.

issue7144
review44251003