Fix Decimal conversion for SQLite
changeset 3de9de3f476c set initial balance value as Decimal. But as the SQLite
Decimal loop break on the first correct instance, it fails to convert other
value coming from the SQL if the default value is tested first.
As there is any way a second loop, it is faster to make the conversion in this
second loop.

issue7284