Register Sale Line only when sale_price_list is activated
The extras_depend is on sale_price_list and not sale. If sale is activated but
not sale_price_list, the module may be loaded before sale and the Sale Line
would not yet be present in the Pool.

issue7391
review47381002
(grafted from cb0b3b9faded6f7b9eebfdd5778939b796e7b341)

--HG--
branch : 4.8