Manage exclude_field as group attribute
We can not rely on the screen linked to the group to provide this information
because if the group is the value of a One2Many inside a One2Many, the screen
is shared between multiple records and so only one record is linked to the
screen. So when the top record is validated all the records not linked to the
screen may be considered invalid because the exclude_field will not be
excluded from the validation.

issue7715
review46651002
(grafted from 59135fd7ced2295c37dc6fd9936c430587114948)

--HG--
branch : 4.2