Search queries inside Select and add different lot instances
When compute_quantities_query is called with with_childs, the selected move
table is a union. So we must search inside this union for the actual move
table.
As we may need to join with lot table multiple times, we must ensure to always
use a different instance of the table to avoid alias clash.

issue7882
review48701002
(grafted from ae31a427c3966a91e3267bdd10bb2ecc1524d9ae)

--HG--
branch : 4.8