Remove call to window_list_toplevels when destroying NoModal
The test was added by changeset 5b5d0c1660e9 to prevent case when the parent
was destoyed before the NoModal. This happened when creating a record from the
search window. But this is no more the case with changeset b87844203f77 as the
search window is destroyed before the creation of the form window.
As the window_list_toplevels has shown multiple times to generate segmentation
fault, it is safer to stop calling it and remove the test which is normally no
more needed.

issue8049
review52621002
(grafted from 57b75ada1af4b716ab162cdbfd8a896e886c936f)

--HG--
branch : 4.6