Do not use set_text of Many2One and Reference when setting only string
changeset 13d81b52b4c4 replaced some direct call to wid_text.set_text by the
set_text but it set_text on Reference expect to receive a tuple (from
get_client). So when set_text is called just to clear or update the wid_text
it loose the selection of Reference.
Indeed as most of the time, it is called with an empty string, there is no
need to set the position of the entry.

issue8114
review54541002
(grafted from 46342207060087e5733df8bddf35418f374a6cf2)

--HG--
branch : 4.6