Remove product_suppliers domain constraint on product
As product suppliers can be defined at the product and template level but also
using a MatchMixin, it is no more flexible to have a domain using it on the
product. It does not hurt the data integrity because if the field is changed
the product_supplier field will also be validated (because it depends on it).
Also the on_change's ensure a default integrity for the user.

issue6799