Do not fail on purchase line not linked to a request
If a line is added to a purchase for drop shipment, it will not be linked to a
purchase request and so there will be no link to any sale line.
In such case, the drop shipment will warn about the new customer move without
origin and the user will be responsible to invoice it manually.

issue8038
review269181002
(grafted from 64126fff63cff950b7d8273aa94923c3a6e4f7a7)

--HG--
branch : 4.4