Ensure to set the same quantity and unit on sale line
The on_change_product calls may change the unit if it is not in the right
category. In such case, we want to raise an error message to the user instead
of silently change the unit and quantity.

issue8079
review279221004
(grafted from 3d1a78feae4f3816a47eb9fb16fc3068679eb57b)

--HG--
branch : 4.8