Update selection to the dropped record
After drag&drop, the treeview keeps the selection at the path where the
drag&drop started. But as the current_record is not changed, this creates an
inconsistant display. So we ensure that the dropped record is selected at the
end.

issue8192
review267371002
(grafted from c3d403305e97551e8a7df465c18d31547abce081)

--HG--
branch : 5.0