Ensure to set the right UoM on purchase line
The UoM of the request must be constrained to the same category as the default
uom of the product.
But also we must ensure that on_change_product does not change the quantity or
unit. The purchase creation should raise an error if the unit is not valid on
the purchase line.

issue8358
review257471002