Set url on switch callback only if it is the current tab
The callback can be triggered when the tab is no more current.
E.g. at the end of a wizard when the screen is reloaded and displayed, the
current record is reset which trigger the callback but the wizard may have
open a new tab meanwhile.

issue8432
review259671002
(grafted from eab82c83ad32ab4483a12fc1b432df17f7e9d2e6)

--HG--
branch : 5.0