Do not use DOM order to find current tab
The DOM and the tabs list are not ensured to have the same order because the
DOM is modified after the view promise is resolved.
To rely only on a single information source, we store the tab as data on the
nav li.

issue8443
review249521004
(grafted from d3d2d9a2cc92cf66097bc3f2b9894e247c01ab14)

--HG--
branch : 4.6