Do not use record in __init__ of form widget
During the initialization of the widget, the view may not yet have a proper
record. But update_img access the record of the view.
Also it is safer for list-form to define the _record before calling super in
case there is an access to record attribute.

issue8616
review285981002
(grafted from 9953ba78be722cf41f66c8750e7c4dfeecf1974c)

--HG--
branch : 5.2