Ensure on_change arguments are loaded
As proteus is lazy about field it read, after a save call there is no field
read. But if the next on_change need the value of a field no loaded, it will
send an empty value because Eval accessors do not trigger a read. So to behave
like other client (which load displayed fields), we ensure to load the needed
arguments.

issue8450
review264411005