Do not re-construct tree view if the same records are to display
Re-construct the tree looses the cursor, the expand state and the editable
cell. Until now, we always re-constructed the tree on display because there was
no simple test like the size for the list.
Now we compare the flat list of records following the expanded row. We
re-construct the table only if they are different.

issue8854
review264671003