Skip eager loading if field is already in the cache
The Many2One and Reference fields are stored in global cache. If they are
already in the cache, the other eager fields should also be in the cache and it
is highly probable that no read will be needed. So it is better to skip the
eager loading.

issue8863
review272541003