Use the number of category for pairing the categories and the lines
The Elegant Pairing generate integer out of range as soon as there are more
than 46340 sale lines (under PostgreSQL).
As the number of category is almost fixed and small, we can pair by multiplying
the line id per the number of category and add the category id.
We must try to use number of category that does not change if new categories
are created meanwhile. For that we always takes the first power of ten bigger
then the current number. This reduces the number of cases when all ids will be
changed because a new category was added.

issue8941
review278321003