Evaluate PYSON domain with the origin record
When a relation domain is not valid, the domain used to set on the exception
must be evaluated with the origin record and not the invalid target record.
We take the first origin record as any of the passed record will generate the
same domain per design.

issue9383
review317531002
(grafted from 728cb4a36144e9f3225e2ff916c132ce82a1f8a7)

--HG--
branch : 5.4