Delete read data after having instantiated all targets
Some read value may be needed to instantiate other records if it is used in the
context.
Also the test to fill fields dictionary was always false as it tested the field
instance instead of the name.

issue9380
review298171002
(grafted from 242491ac0135a2c4d4f186313a5b88aaa2966277)

--HG--
branch : 5.4