Use editable path to update editable on display
The current record is not necessary linked to the editable. For example when a
new record is added, the current record is changed before the editing is done.
Also the CellRendererCombo does not emit remove-widget when focus is changed
which prevent to clear the editable variables.

issue9618
issue9266
review314471002
(grafted from 07b4d22509fcbabb19118233e1cacfb4f15d5f3a)

--HG--
branch : 5.4