Catch any exception for server version RPC
It is the first method called with the input of the user.
The login window does not support that it raises an exception and as a lot of
different exceptions can be raised (and may change depending of Python
version), we must catch all of them. It does not hide any programming error as
they are logged.

issue9408
review292561004
(grafted from feb8125c72d23ae3d5aab49f2745ed25412d2c02)

--HG--
branch : 5.4