Return false when modified is not saved and not created
The return of modified_save method is used to decide to execute an action or
not. When a new record is cancelled, the client select the previous record so
we must not run the action on this one.

issue9881
review337111002
(grafted from bfd7ae2788343d6562924a38f66a480ae7b8ad0c)

--HG--
branch : 5.0