Fully set product domain of move on setup
When modify in __setup__ the domain of a field defined in the same class, this
alter the original copy of the field. So the domain is appended to itself on
each __setup__ call (each time the model is extended by another module).

issue10406
review336241002
(grafted from 3670b1a582ecf41612f1f84b73b8ab68f8bf8b1b)

--HG--
branch : 6.0