Do not display search_count if it is not higher than limit
On small set of record when records are deleted the number of record could
become lower than the initial search_count. As maintaining a correct
search_count is complex, it is simpler to never display it if it is lower than
the search limit.

issue10789
review383501003
(grafted from 6b2e33b1025568f56aa23972fc02115da18496fc)

--HG--
branch : 5.8