Remove SignalEvent usage
We register on the screen the windows and on the group the screens to which
events must be sent.

The window of Screen can be a tab, a dialog or even a widget like for One2Many.
They must just have a method with the proper name to receive the corresponding
events.
The group_list_changed event is dedicated to synchronisation of the Gtk
TreeView and ListBox models.

issue11059
review395151002