Round quantity computing in search clause
The internal quantity is stored without precision constraint and so the sum
result of many rows cumulate the rounding issue.
So the quantity must be rounded to return correct result when operator is
equality.
But as getting the digits from the product default uom would require to make
many joins, the simplest and more efficient solution is to use the highest
digits of the units.

issue11235
review390001002