Remove user login and user preferences cache
The user login cache could result in an error when using trytond in a
multiprocess environment due to a delay in the broadcast of the cache
invalidation.

The user preferences cache was there only to set the context when validating
the records but it is useless since the context is now set from the
transaction.

review401251002
issue11430