Do not set focus on date when input changed for Firefox
Firefox fires a change event when user use the navigation per month and year.
So we can not set the focus on the date entry without closing the input picker.
In order to navigate with tab after picker is closed, we must ensure to not
have the focus on the icon.

issue11715
review426031003
(grafted from 229609bd10badedef59ac60b4e4dd6ed8764f288)

--HG--
branch : 6.4