Assign supplied outgoing moves instead of inventory moves
If the output location is the same as the storage location, there is no
inventory moves on the outgoing shipment, so the outgoing moves must be
assigned.
If the input location is the same as the storage location, there is neither
inventory moves on the incoming shipment, so the quantity must be computed
based on the incoming moves.

Closes #11972