Instantiate always Screen for xxx2Many fields with the same parameters
The screen must be instantiated with the context of the parent screen so
eventual field_view_get would give the same result.
The limit must always be set to None/null to ensure to load all the records.
The breadcrumb must be set to show navigation to the user.

Closes #3140
(grafted from ced0956c41cd1192b98eb4a9e9e23f85c86b6240)

--HG--
branch : 6.6