Simplify stock forecast complete wizard
It set the context on the products selection so the user can see the quantity
that will be set. As it is possible to search on this quantity, there is no
more the need to prefill the selection with products having a quantity.
Also there is no need to have two steps for the wizard as the context is
computed dynamically.