Set order for stock moves on shipments
On shipment a good default order is by target location and product, but we add
the id to have deterministic result and remove the target location if by
design it is always the same.
On report with sequence on location, the order is fixed by the sequence of the
location.

Closes #7819