Restore scroll position to the position before loading more rows
When the scroll is at the bottom (and a footer is present), when new rows are
added the scroll stay at the bottom which trigger again to load more rows.
This continue until reaching the end of the tree.
To prevent this behavior we restore the position of the scroll using the
height from the top.

Closes #12223