Clear cache before commit each module
As the cache is rollback to avoid dead lock, the transaction commit could
restore cache version that is no more valid.
For example the language cache of `ir.translation` which is filled initially
with only English and then reset when other languages are created by the `ir`
module. So at the end of `ir` module load, the cache is reset to contain only
English as it was the content before the first cache reset. This prevents to
load any other language from other modules.
So to ensure the next module is loaded using a fresh cache, all the caches are
cleared.

Closes #12246