Parse authorization header for session scheme without token
Since Werkzeug >= 2.3, it always tries to parse the Authorization header as a
digest. If the session value which is encoded in base64 needs padding, the
trailing `=` are interpreted as header values and the token attribute is not
filled.
In this case we must parse ourself the Authorization header to decode the
base64 value as session.

Closes #12259
(grafted from 0bd0b8a0e20ad8d2f3d13cc8d879c78717817b7a)

--HG--
branch : 6.0