Do not validate model that requires transaction lock
Since c8771e45b861 an exception is raised if the table was not locked on
transaction starts.
When validating some records with trytond-admin, we do not want to restart the
validation process to lock the tables to avoid creating conflict with the
running server.
(grafted from ee4a21126b3a21ee5e10daedff7ed3f12c088cf7)

--HG--
branch : 6.8