Clear caches after commit when installing modules
A call to clear_all is required after the commit in order to ensure that the
transaction uses an empty cache from this point. clear_all is not required to
prevent the deadlock on ir.cache because the rollback ensures that nothing
changed from the point of view of this transaction.

Close #12343
(grafted from 48396d9cfed6caccb3fca6d981fec5016a1f6ca2)

--HG--
branch : 6.8