Cache the url for authentication service
When starting the desktop client you can select a profile. And by doing
that, the client sends a request to fetch the authentication services.
The data of the request is then cached but the url is not so the client
forget about it. Because of the cached data, when selecting a particular
profile data is taken from the cache but the url is not updated leaving
it unchanged and empty or with a url from last request.

Closes #12440