Move sleep loop introduced in a32b84d96649 to prevent deadlock to tests
In this changeset we fix an issue with testing.

When using the listener for cache invalidation it results into the database
lock used by the pool that is acquired twice. Once when calling Pool.init ;
once when the transaction is started which calls Cache.sync which in turn calls
Pool.refresh and finaly Pool.stop. It tries to acquire again the re-entrant
lock (but it cannot because it's another thread).

Closes #12457
(grafted from 3b618f310b354126e4dbb0f19ad139c5bf879b9e)

--HG--
branch : 6.8