Deduce sold shipment cost from shipment once sale is shipped
When the shipment cost method is by order, the sold shipment cost is
dispatched between shipments once the sale is shipped.
The amount is stored on the shipment for each sale (in case the shipments have
been grouped) and deduced from the shipment cost when it is set on the stock
moves.

Closes #12506