Select only SEPA mandates of the same company as the payment
Since ce8d7a19766a record rules (and especially company rules) are no more
enforced without check access. So the mandate selection should test explicitly
for the same company.
For that a new generator method is added to party to yield mandates that can
be used for a payment so this method can be easily extended to add more
constraints.

Closes #12324