Set visibility of active button when screen is displayed
Setting it on the domain parser property does not ensure that it is always set
if the parser has been cached or the context screen is not valid.
Also it is not right that accessing a property trigger a change of state.
As the button visibility can change between views, the active domain must only
be added if the button is currently visible.

Closes #12549