Fallback to original copied value of assigned stock move
The default values used by 9d4aff3c72e2 to group copy calls may contains keys
that are not in the modified values because of the union.
In order to keep the copy call grouped as much, we just need to fallback to
the original data value for the fields that are not modified.

Closes #12290