Update trees of chart of accounts with 2 passes
When a new parent is added to a tree template like types, taxes or tax codes,
the corresponding record of the new template parent is not yet created.
So we need to use a default value for the parent and update again the tree
after creating the new records.

Closes #12755
(grafted from 561f9c7dbdc2a2c39c1fe5f6bf7274d01ee9ebd6)

--HG--
branch : 6.0