Use empty value Function xxx2Many fields without initial value
Since d142b33fa3ab the on_change_with relies on _changed_value. But for
Function xxx2Many fields if their value is not in the initial value, we should
not retrieve it from the database because the result will be the same as the
one set by on_change_with.

Closes #12836
(grafted from 13d2870d43c833a8629b451edd82ac2efdc27009)

--HG--
branch : 7.0