Handle SOAP errors in EU Check Vies service
When introducing CheckVIES wizard in 7c90d7100f10, it was based on module
'vatnumbers' which returned an XML tree where we needed to pick the
'faultstring' from.
Since aeb2d1398456 it uses 'stdnum', which uses either 'zeep', 'suds' or
'pysimplesoap' as backend, all of which provide the faultstring in the
exception arguments. Unfortunately 'suds' puts it into 'args[1]', thus a loop
is required.

Also add missing error-codes from EU Check Vies service.

Closes #11147